[BSK-1729] fixing a few more gcc linux compiler warnings
Former-commit-id: 63b8752472a29a328e10a41a1ed7db1270a95c35
Former-commit-id: 37f804c9846d72e1f05afa35df950ce008421883

+++ src/fswAlgorithms/_cModuleTemplateFolder/cModuleTemplate/cModuleTemplate.c
@@ -97,7 +97,7 @@ void Update_cModuleTemplate(cModuleTemplateConfig *configData, uint64_t callTime
     /* this logging statement is not typically required.  It is done here to see in the
      quick-start guide which module is being executed */
     char info[MAX_LOGGING_LENGTH];
-    sprintf(info, "C Module ID %lld ran Update at %fs", moduleID, (double) callTime/(1e9));
+    sprintf(info, "C Module ID %lld ran Update at %fs", (long long int) moduleID, (double) callTime/(1e9));
     _bskLog(configData->bskLogger, BSK_INFORMATION, info);
 
     return;
